Changed Displacement to a float

+++ MojioSDK/mojiosdksrc/src/main/java/com/mojio/mojiosdk/models/Engine.java
@@ -6,7 +6,7 @@ package com.mojio.mojiosdk.models;
 public class Engine {
     public String Name;
     public int Cylinders;
-    public int Displacement;
+    public float Displacement;
     public String MaxHp;
     public String MaxHpAt;
 }
